requirements.txt
aiohttp
APScheduler
astroid
astunparse
async-timeout
attrs
braceexpand
certifi
chardet==3.0.4
colorama==0.4.3
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d3adc860957ccef05baeec2e2dd8
discord.py==1.4.1
humanize==2.6.0
idna==2.10
import-expression==1.1.4
isort==5.5.1
jishaku==1.19.1.200
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
psycopg2==2.8.6
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
grequests